Removed all code issues porting from iOS 2 to iOS 7

+++ Class/GalleryDatabaseDelegate.m
@@ -88,7 +88,7 @@
 				
 				
                 GalleryItem *galleryItem = [[GalleryItem alloc] init];
-                [galleryItem setUniqueId:aUid];
+                [galleryItem setUniqueId:(int)aUid];
                 [galleryItem setImageName:aImageName];
                 [galleryItem setTitle:aTitle];
                 [galleryItem setMethod:aMethod];
